+++ tools/quake3/q3map2/minimap.cpp
@@ -146,8 +146,8 @@ static void MiniMapRandomlySupersampled( int y ){
 	int x, i;
 	float *p = &minimap.data1f[y * minimap.width];
 	float ymin = minimap.mins[1] + minimap.size[1] * ( y / (float) minimap.height );
-	float dx   =                   minimap.size[0]      / (float) minimap.width;
-	float dy   =                   minimap.size[1]      / (float) minimap.height;
+	float dx   =                   minimap.size[0]       / (float) minimap.width;
+	float dy   =                   minimap.size[1]       / (float) minimap.height;
 	float uv[2];
 	float thisval;
 
@@ -174,8 +174,8 @@ static void MiniMapSupersampled( int y ){
 	int x, i;
 	float *p = &minimap.data1f[y * minimap.width];
 	float ymin = minimap.mins[1] + minimap.size[1] * ( y / (float) minimap.height );
-	float dx   =                   minimap.size[0]      / (float) minimap.width;
-	float dy   =                   minimap.size[1]      / (float) minimap.height;
+	float dx   =                   minimap.size[0]       / (float) minimap.width;
+	float dy   =                   minimap.size[1]       / (float) minimap.height;
 
 	for ( x = 0; x < minimap.width; ++x )
 	{
@@ -198,11 +198,11 @@ static void MiniMapSupersampled( int y ){
 static void MiniMapNoSupersampling( int y ){
 	int x;
 	float *p = &minimap.data1f[y * minimap.width];
-	float ymin = minimap.mins[1] + minimap.size[1] * ( ( y + 0.5 ) / (float) minimap.height );
+	float ymin = minimap.mins[1] + minimap.size[1] * ( ( y + 0.5 ) / minimap.height );
 
 	for ( x = 0; x < minimap.width; ++x )
 	{
-		float xmin = minimap.mins[0] + minimap.size[0] * ( ( x + 0.5 ) / (float) minimap.width );
+		float xmin = minimap.mins[0] + minimap.size[0] * ( ( x + 0.5 ) / minimap.width );
 		*p++ = MiniMapSample( xmin, ymin ) / minimap.size[2];
 	}
 }
